Adding the Catalyst dialect to the Python compiler (#7901)
**Context:**
This PR ports the catalyst dialect from Catalyst over to the Python
compiler.

**Description of the Change:**
It does so by following these steps:

- [x] Dumped a json format of the Catalyst dialect tablegen using the
`llvm-tblgen`.
- [x] Deleted `“Builtin_Dialect”` from
`"Dialect":["Builtin_Dialect","Catalyst_Dialect"]` within the json file.
Otherwise the next step would fail.
- [x] run `xdsl-tblgen` on the modified json to get the dialect python
file.
- [x] Sometimes xdsl parses attributes as properties and expects the
property to be included in the declarative assembly format. However,
this is something in progress within MLIR and not all dialects support
properties yet. we either need to manually change properties to
attributes or use this option`irdl_options = [ParsePropInAttrDict()]`.
the latter was done here.
- [x] MLIR has a trait called `TypesMatchWith` which allows mathing
between two types. This I think is not currently supported in xdsl.
There are ways to manually assign types but this didn't work in the case
of matching a type with type of elements of a list (see
[here](https://github.com/PennyLaneAI/catalyst/blob/c54b13229525eae6336012ac96598623c1337451/mlir/include/Catalyst/IR/CatalystOps.td#L41-L58)
for how it is used in Catalyst dialect in cpp). We resolved the issue by
marking `ArrayListType` generic and also replacing the `BaseAttr()`
method which fails to infer the elemtn types of the list with `base()`.
`base()` calls another method called `irdl_to_attr_constraint(irdl)`
which recursively looks for the type of arguments of the original
attribute if it is parameterized and generic (see
[here](https://github.com/xdslproject/xdsl/blob/cb8d568ec34caef1a79517d0892dabc4870c1175/xdsl/irdl/attributes.py#L491C5-L526C10))

- [x] Simplified or improved some types miscategorized or generalized in
the automatic script.

- [x] Added tests for the dialect

**Benefits:**
one of the main dialects in Catalyst is added to the python compiler.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-93613]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>