[Decomposition] Register work-wire usage of decomposition rules (#7861)
**Context:**
With dynamic qubit allocation, whether a decomposition rule can be
performed may depend on whether there is enough work wires available.
The number of work wires that a decomposition rule uses should be
registered as a resource requirement.

**Description of the Change:**
~Implements a `require_work_wires` decorator to be applied on
decomposition rules to declare the number of work wires that this
decomposition rule intends to use.~
Adds a `work_wires` argument to `register_resources`.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-94399]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>