Default implementation of setup_execution_config uses "device" mcm (#7968)
**Context:**

**Description of the Change:**
Update the default implementation of `Device.setup_execution_config` to
use the `"device"` MCM method if `"device"` is available, regardless of
finite shots or not.

**Benefits:**
Aligns with the default strategy discussed.

**Possible Drawbacks:**
The only devices affected by this are one that defines a `TOML` file
with `"device"` included in its mcm methods, and does not override
`preprocess`, which is none of the devices.

**Related GitHub Issues:**
[sc-95418]