Small refactor of `queue` method for `ResourceOperators` (#7974)
**Context:**
Resource operators are tracked using the queuing architecture. The
default `queue()` method adds an instance of the Resource operator to
the active queuing context.

In some cases operators consume other operators and thus those inputs
need to be removed from the queue. This responsibility was being
absorbed into the `queue()` leading to diverging implementations for the
various Resource operators depending on their inputs. In this PR we move
that responsibility to a more general, separate method called `dequeue`.

**Description of the Change:**
- Refactored the diverging `queue()` method implementations 
- Add a new `dequeue()` method to remove from the queue.

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>