Add PrepSelPrep to new decomposition framework (#7385)
**Context:**

`PrepSelPrep` is a testing ground for developing a new way of handling
work wires. We are also tailoring that new way toward qfuncs and the new
decomposition structure. So to optimally use it as a testing ground, we
need some ops to be ported over.

**Description of the Change:**

Add decompositions for:
* `Prod`
* `StatePrep`
* `MottononenStatePrep`
* `PrepSelPrep`
* `Select`

**Benefits:**

More coverage for new decomposition structure.

**Possible Drawbacks:**

**Related Shortcut Stories:**

[sc-90542] [sc-89612]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: dwierichs <david.wierichs@xanadu.ai>