xDSL diagonalize_measurements (#7864)
Notes:

- rather than showing a useful error message if the program contains
non-commuting observables, it currently just returns incorrect results;
this shouldn't occur as long as we apply `split_non_commuting` first, so
we aren't worrying about it for now.
- we are currently inserting the diagonalizing gates directly before the
observables, which means that the transform breaks the (maybe not really
established in MLIR) convention of having an "ops" section followed by
an "observables and measurements" section. We will come back to this and
fix it once the framework for separating state transformation and
readout for split_non_commuting is complete.
--------------------

**Context:**
A python compiler (currently xDSL) implementation of the
`diagonalize_measurements` transform is an important step in enabling
current MBQC workflows via the Catalyst pipeline.

**Description of the Change:**
We add a simple implementation of the diagonalize_measurements
transform. It diagonalizes X, Y and Hadamard.

**Benefits:**
We can diagonalize measurements at the MLIR stage of compilation

**Possible Drawbacks:**
This is not a full implementation of the functionality of the tape
transform `diagonalize_measurements`. It implements only the basic
functionality of diagonalizing circuits with observables X, Y, Z, and
Identity to create circuits with only Z and Identity observables. For
example, it doesn't allow diagonalization of only a subset of `{X, Y}`,
and doesn't allow for diagonalization of `Hadamard` at all.

[sc-92510]

---------

Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>
Co-authored-by: Shuli Shu <31480676+multiphaseCFD@users.noreply.github.com>