Adding the `compute_depth` kwarg to `qml.specs` (#7998)
**Context:** The depth computation in `qml.specs` has been identified as
a bottleneck in performance, and users are often more concerned with
other information provided by `qml.specs`.
The depth computation should be accessible elsewhere, or at least be
made optional.

**Description of the Change:**

`qml.specs` receives an additional kwarg: `compute_depth=True/False`.
This makes the expensive depth computation performed by specs optional.
The default is `compute_depth=True`, making depth computation ON by
default (as it was before - no change in default behavior).

Notice that `qml.specs` calls a series of functions/methods, which
eventually call the `specs` property of `QuantumScript`. This is a
property, so it cannot receive arguments. This implies that we cannot
provide further details about how to compute the `specs` (for example,
as in this case, which specs we want to compute or not compute) unless
we modify this class.

This is a very high-level workflow logic of `qml.specs`:

`qml.specs` ---> `QuantumScripts.specs` ---> `_count_resources` --->
`Graph.get_depth`

The simplest solution is to add a `compute_specs` method to the
`QuantumScript` class and move the computation logic here. This provides
much more flexibility for computing specs of a given tape. In this case,
we use this new flexibility to skip the computation of the `depth`. We
added a test both for `qml.specs` and the new `compute_specs` method of
`QuantumScript`.

**Benefits:** The depth spec can be skipped if required.

**Possible Drawbacks:** None that I can think of.

**Related GitHub Issues:** None.

[sc-96639]