Simplify the UI of the `HilbertSchmidt` and `LocalHilbertSchmidt` templates (#7933)
**Context:**

The purpose of this PR is to remove the dependency on `QuantumScript`
from the signature and metadata (hyper)parameters of the
`HilbertSchmidt` and `LocalHilbertSchmidt` templates. Furthermore, the
goal has been expanded to simplify the UI of this template considerably.

**Description of the Change:**

- The `u_tape` argument in the `HilbertSchmidt` and
`LocalHilbertSchmidt` templates becomes `U`, and it can be a single
operation or an iterable of operations
- The `v_function` becomes `V`, and it can be a single operation or an
iterable of operations
- the `v_wires` and `v_params` arguments are removed and they are
inferred from `V`

Before the way to use the template was:
```
 qml.HilbertSchmidt(v_params, v_function=v_function, v_wires=v_wires, u=u_ops)
```

Now the UI has become much simpler:

```
 qml.HilbertSchmidt(V, U)
```

**Benefits:** The signature of the function is much simpler and does no
longer depends on a tape.

**Possible Drawbacks:** This is a breaking change, but this template is
probably not heavily used.

**Related GitHub Issues:** None.

[sc-95625]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>