Do not to try to capture loops that should not be entered (#8026)
------------------------------------------------------------------------------------------------------------

**Context:** We were erroring out when trying to capture loops with a
`start=stop=0` when the iterator was used as a index into a jax array
with length 0. Since there existed no value of the index for which the
array's entry at that index existed, it was just not something we could
capture.

**Description of the Change:** When we have a loop with start=stop, we
just don't try to capture it. It is a NOP. We execute the version of the
decorated function that we usually do when capture is disabled, which
simply does not enter the loop and returns the unmutated args to the
decorated function.

**Benefits:** Better developer experience... loops with start=stop no
longer entered (that was confusing!)

**Possible Drawbacks:** 

**Related ShortCut Stories:** [sc-97074]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>