Add a no_analytic transform (#8037)
**Context:**

We have `no_sampling` as a transform to ensure compatibility with
certain execution pathways that only work with analytic execution, but
we don't have a `no_analytic` transform, which is relevant for hardware
devices.

To provide this functionality to date, we've been validating that
`device.shots != None` in device `__init__`s, instead of validating the
tape shots in the transform program. This is no longer the direction we
want to go, as we are deprecating `device.shots`, and also isn't the
ideal place for validation because calls to the QNode can override
`device.shots`.

**Description of the Change:**

We add a `no_analytic` transform

**Benefits:**
It's better practice, and it supports the deprecation of `device.shots`
in plugins.

[sc-97221]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>