Ftqc device frontend (#7656)
**Context:**
We want to prototype executing on a backend with a feedback loop
(execution of part of the tape, updates based on measurement results,
continued execution), while leaning where possible on the existing
architecture in PennyLane.

We already have tape transforms in place for conversion to the MBQC
formalism, as well as the framework for a transform for using a Pauli
tracker to consolidate corrections.

**Description of the Change:**
- We create an `ftqc.device` whose main functions are to:
1. define a `preprocess` that is essentially the compilation pipeline
for going from an arbitrary circuit defined in PennyLane to the MBQC
formalism
2. define a setup for the execution config with the relevant entries for
integration with PennyLane's execution and device API
3. define its own execution function that will handle the overhead for
executing the tapes stepwise with corrections

In this PR, only the first two points are relevant. The 3rd will be its
own PR.

- We create backends that are currently just wrappers for
"lightning.qubit" and "null.qubit", but with their own `capabilities`
that reflect the limitations of the fake HW device.
- In a follow-up PR, the `lightning.qubit` backend will be updated with
modified `simulate` and `execute` methods that allow us to maintain the
`_statevector` from one execution to the next, unless we are explicitly
resetting the system. The `null.qubit` backend may be updated if needed
to mimic a mocked version of the same behaviour, but likely won't need
much modification.

**Benefits:**
The main goal here was to create two separate objects with different
responsibilities for the execution pipeline - one that applies
preprocessing and orchestrates the feedback loop, and another that
executes. Which thing goes in which box is subject to change as this
prototype develops over this iteration.

The important thing here is that the boxes exist, the preprocessing
works as expected, and we can use this setup to execute arbitrary
PennyLane circuits in the simplest (online corrections after every
operation) implementation of textbook MBQC.

[sc-90413] [sc-93075]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>