[Capture] `cond` can accept an operator `true_fn` with no `false_fn` provided (#8060)
**Context:**

One thing both core pennylane and current catalyst both support is:
```
qml.cond(True, qml.X)(0)
```

Capture currently forbids this, because the `true_fn` and `false_fn`
have different outputs.

If we make the assumption that the above syntax always means "queue an X
operator" instead of "return an X operator", we can still consider it
well defined.

So if no `false_fn` is provided and the input is a subclass of operator,
we should just assume that the operator should be "queued" not
necessarily returned.

**Description of the Change:**

Removes the output from `true_fn` if it's a subclass of `Operator` and
no `false_fn` is provided.

**Benefits:**

Smoother change to using program capture as the new frontend.

**Possible Drawbacks:**

Someone might actually want the operator instance. This is more of a
"patch" for a specific use case than generic behavior.

**Related GitHub Issues:**

[sc-97433]