Add `xdsl_extras` module to Python compiler for functionality not available upstream in xDSL (#8067)
**Context:** In MLIR, we are able to constrain the rank of a tensor
using TableGen directives, for example

```
AnyTypeOf<[
    1DTensorOf<[Complex<F64>]>, MemRefRankOf<[Complex<F64>], [1]>
]>:$in_value
```

which defines a value `in_value` and constrains its type to be either a
complex<f64> tensor or memref of rank 1. The equivalent functionality is
currently not available in xDSL. For the moment, if we want to apply
such constraints, we must define these attributes ourselves.

**Description of the Change:** Adds a new module
`pennylane.compiler.python_compiler.xdsl_extras` containing two new
constraint attributes, `MemRefRankConstraint` and
`TensorRankConstraint`, which can be used to constrain the rank of a
memref or tensor type, respectively.

**Example:**

These new constraint attributes can be used as follows:

```python
in_value = operand_def(
    (TensorType.constr(i64) & TensorRankConstraint(1))
    | (MemRefType.constr(i64) & MemRefRankConstraint(1))
)
```

which constrains the type of `in_value` to be either a rank-1 tensor or
memref with type i64.

**Benefits:** Better constraints on our IR.

**Possible Drawbacks:** Eventually these constraints may be upstreamed
to xDSL, in which case they should be removed from the PennyLane Python
compiler.

[sc-97540]