Make `assign_states` public, faster and remove a bug (#7880)
**Context:**
In
`pennylane.templates.state_preparations.superposition.py::_assign_states`,
some computational basis states are mapped to new computational basis
states.
It is claimed that states that are both in the set of inputs and the set
of outputs are fixed points of the produced map, but this is not true.

`_assign_states` is used in `qualtran_io.py` but is a private function,
so[ it might be better to make it
public.](https://github.com/PennyLaneAI/pennylane/pull/7866/files#r2205072082)

**Description of the Change:**
- Make the mentioned states fixed points of the produced map
- Speed up the function by looping over the set of states fewer times
- Make `assign_states` public-facing.

**Benefits:**
Better import hygiene, faster code, one bug less.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>