Optimize `BasisRotation` for real-valued inputs and extend theory in docstring (#7765)
**Context:**
`qml.BasisRotation` implements a basis rotation in the space of
orbitals, mapped to quantum circuits.
In general, the decomposition contains `PhaseShift` and
`SingleExcitation` gates.
If the rotation matrix is real-valued, and thus orthogonal, the
decomposition can be simplified to Givens rotations (`SingleExcitation`)
only (for special orthogonal, i.e., determinant $1$), or Givens
rotations and a single `PhaseShift` (determinant $-1$).
In addition, we recently gained some insight in how to understand the
mapping from orbital rotations to qubits.

**Description of the Change:**
Add a branch to the decomposition of `qml.BasisRotation` that is
triggered if the rotation matrix has a real number data type. It uses
zero or one `PhaseShift` operations instead of up to $(N^2+N)/2$ phase
shifts, and the same number of `SingleExcitation` gates as the
complex-valued decomposition.

The second major change is an extension of the theory part of the
docstring of `BasisRotation`, including a Lie algebraic explanation for
the mapping.

**Benefits:**
Cheaper decomposition. More math context in docs.

**Possible Drawbacks:**
More complex decomposition logic.

**Related GitHub Issues:**

[sc-94397]

**To do**
- [x] First docstring pass
- [x] Second docstring pass
- [x] test completion (matrix tests?)
- [x] reduce logic if possible
- [x] changelog

---------

Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: Pablo Antonio Moreno Casares <paulmorecasa@gmail.com>