No longer squeeze singleton dimension in samples (#7944)
**Context:**

Similar to #7917 , but doesn't squeeze observables.

Currently, we always squeeze out singleton dimension from `qml.sample`
measurments. This leads to lots of additional logical branches, and
error-prone postprocessing. It can be really easy to loose track of
which dimensions are being sliced into.

Catalyst also never squeezes out singleton dimensions. If we want to
unify the behavior of pennylane and catalyst, we need to either start
squeezing in catalyst, or stop squeezing in pennylane. We are opting to
stop squeezing in pennylane.

**Description of the Change:**

Switches to that sampling wires *always* returns an array of shape
`(shots, wires)`, regardless of whether or not we have a single shot or
a single wire.

**Benefits:**

Much simpler, more robust code and unified behavior between pennylane
and catalyst.

**Possible Drawbacks:**

This is going to be a breaking change, and there's really no way to
soften this to a deprecation. We just have to pull the bandaid off.

Need an accompanying issue to lightning.

**Related GitHub Issues:**

[[sc-95797](https://app.shortcut.com/xanaduai/story/95797)]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>