Add `partial` keyword argument to `Select` and simplify its decomposition if `partial=True` (#7658)
**Context:**
`qml.Select` can be reduced if two conditions are met:
1. It is a *partial Select* operator, applying fewer operators ($K$)
than would be maximally possible for the given control wires ($2^c$ for
$c$ control wires)
2. the control wires state does not have overlap with computational
basis states that are not used by `Select`.

The first condition can be checked based on the inputs `ops` and
`control` to `Select`.
It is in general not possible to figure out on the fly whether the
second condition is satisfied. Instead, a new keyword argument `partial`
signals to the operator that it may assume this second condition,
allowing the decomposition to reduce itself.

The simplification rule is detailed in a [compilation
page](https://pennylane.ai/compilation/partial-select).

**Description of the Change:**
- Add the new keyword argument `partial` that signals that the second
condition above is met.
- Adapt the multi-control decomposition of `Select` to make use of the
reduction if the condition is met. (Condition 1 is checked based on the
input operators).

(If there are $K=2^{c}$ operators, there is no difference.)

**Benefits:**
Cheaper decomposition of `Select`

**Possible Drawbacks:**
new keyword argument

**Related GitHub Issues:**
This will be helpful to develop a systematic approach together with
unary iterator decomposition in #7623 .

[sc-94728]

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>