Temporarily xfail lightning test (#8130)
**Context:**

A change to `Select`'s decomposition broke lightning. We responded by
natively supporting `Exp` and `SProd` on lightning. This has broken
`test_exp.py`.

**Description of the Change:**

To unblock the CI, we are xfailing that test while we get a fix.

**Benefits:**

**Possible Drawbacks:**

We aren't actually solving the problem. Just letting other work proceed
while we fix things.

**Related GitHub Issues:**