feat: autograph: support in-place array updates with operators (`*=`, `-=`, ...) (#8076)
## Context

This is an in-direct follow-up PR to
https://github.com/PennyLaneAI/pennylane/pull/8027.

Now that we have index assignment. We wish to add in-place array updates
using pythonic notation.

## Description of the Change

This PR directly mirrors the changes made to Catalyst
[here](https://github.com/PennyLaneAI/catalyst/pull/1143). Basically,
the gist is that we added the
`SingleIndexArrayOperatorUpdateTransformer` class that is able to use
autograph to properly convert any detected array update in our AST.

## Benefits

This allows us to do things like,

```python
import pennylane as qml

qml.capture.enable()

from pennylane.capture import make_plxpr

import jax
import jax.numpy as jnp

def f():
    # setup my_list = [0, 1]
    my_list = jnp.empty(2, dtype=int)
    for i in range(2):
        my_list[i] = i
    # increment second element by ten
    my_list[1] += 10
    return my_list

>>> plxpr = make_plxpr(f)()
>>> jax.core.eval_jaxpr(plxpr.jaxpr, [])
[Array([0, 11], dtype=int32)]
```

## Possible Drawbacks

None identified.

[sc-76134]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>