Extend the StableHLO dialect in the Python Compiler (#8036)
**Context:**
Currently StableHLO dialect in xdsl contains a small subset of
operations available in the OpenXLA repo. A full list of categorized
operations can be found
[here](https://github.com/openxla/stablehlo/blob/main/docs/interpreter_status.md#stablehlo-ops-categories)

**Description of the Change:**
Creates a local shim which mirrors the existing upstream operations as
well as allowing us to extend the stablehlo to include more operation.

**Benefits:**

**Possible Drawbacks:**
This PR should be upstreamed once the StableHLO migration from xdsl is
complete

**Related GitHub Issues:**