Implement strategy for smart shots distribution (#7988)
When estimating the expectation value of a Hamiltonian written as a
linear combination of several terms, we usually need to partition it
into groups of mutually-commuting terms which can be measured
simultaneously. Given a fixed total number of measurement shots, a
crucial task is to determine the optimal allocation of these shots
across the measurements for each commuting group. The goal of this
strategy is to finally combine the individual expectation values to
produce the best estimate of the whole original Hamiltonian. For
example, we might want to weight the number of shots for each group
based on the coefficients of its terms.

The aim of this PR is to set the ground for this "smart" shots
distribution and to integrate this strategy into the already existing
`split_non_commuting` transform. Here we focus on the simple case of a
circuit measuring the expectation value of a single Hamiltonian, but the
idea can be generalized for multiple Hamiltonians as well.

Let's show an example considering the following case:
```python
import pennylane as qml

coeffs = [10, 0.1, 20, 100, 0.2]
observables = [
    qml.X(0) @ qml.Y(1),
    qml.Z(0) @ qml.Z(2),
    qml.Y(1),
    qml.X(1) @ qml.X(2),
    qml.Z(0) @ qml.Z(1) @ qml.Z(2)
]
shots = 10000

ham = qml.Hamiltonian(coeffs=coeffs, observables=observables)
tape = qml.tape.QuantumScript(measurements=[qml.expval(ham)], shots=shots)
```
We want to allow the user to pass a custom function to the
`split_non_commuting` transform in order to define a `shot_distribution`
strategy. For example:
```python
from pennylane.transforms import split_non_commuting

def my_shot_distribution(total_shots, coeffs_per_group):
    # define weighted sampling
    sum_per_group = np.array([np.sum(np.abs(coeffs)) for coeffs in coeffs_per_group])
    prob_shots = sum_per_group / np.sum(sum_per_group)
    return total_shots * prob_shots

new_tapes, _ = split_non_commuting(tape, shot_distribution=my_shot_distribution)

for new_tape in new_tapes:
    print(new_tape.measurements, new_tape.shots)
```
```
[expval(X(0) @ Y(1)), expval(Y(1))] Shots(total=2302)
[expval(Z(0) @ Z(2)), expval(Z(0) @ Z(1) @ Z(2))] Shots(total=23)
[expval(X(1) @ X(2))] Shots(total=7674)
```

[sc-96319]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>