feat: freeze `dict` fields of `ExecutionConfig` using new custom `FrozenMapping` object (#8046)
**Context:**

Currently we can mutate the dict fields of the `ExecutionConfig`,

```python
import pennylane as qml
from pennylane.devices.execution_config import ExecutionConfig  

config = ExecutionConfig(
    device_options={"hi": "bye"}, gradient_keyword_arguments={"foo": "bar"}
)
config.device_options["hi"] = "hello" # Should not be able to mutate this!
```
This is bad.

**Description of the Change:**

Creates a new `FrozenMapping` object that prevents this.

**Benefits:** Fully immutable execution config. This ensures our
configuration isn't being mutated along the way which could lead to
unexpected results. Furthermore, this helps out with the upgrade to jax
0.7.0 as it is now hashable.

**Possible Drawbacks:** Might impact ecosystem?

[sc-96529]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>