Decomposition rule (#8070)
------------------------------------------------------------------------------------------------------------

**Context:**
We would like to provide a way for users to specify that they are using
a compute/uncompute pattern [U, V, U†] so that it may be efficiently
controlled by applying controls to the "middle" operator V instead of
the entire pattern.

**Description of the Change:**
Adds a decomposition rule and registers it under "C(ChangeOpBasis)".
This rule applies ctrl only to the middle op in a compute/uncompute
pattern.
<img width="283" height="57" alt="image"
src="https://github.com/user-attachments/assets/228b8a09-3b56-47bd-a014-d41667e0cb98"
/>

**Benefits:** We can now decompose controlled compute/uncompute patterns
efficiently.

**Possible Drawbacks:**

**Related ShortCut Stories:** [sc-97007]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: ANT0N <39093564+AntonNI8@users.noreply.github.com>
Co-authored-by: Austin Huang <austin.huang@xanadu.ai>