chore: autograph: remove `autograph` kwarg from `QNode` constructor (#8104)
## Context

Both `qjit` and `qnode` currently have an `autograph` kwarg (with
default values False and True respectively).

The solution is that we will use the disable context manager and
`run_autograph` for now, and remove the `autograph` kwarg from the
`QNode` in favour of only the `qjit` decorator and the context managers.

## Description of the Change

Basically just reverts the changes that were made when this kwarg was
added (https://github.com/PennyLaneAI/pennylane/pull/6837).

## Benefits

Aligned with our approach to unify PL and Catalyst.

## Possible Drawbacks

None identified.

[sc-92903]