On-Board pennylane to using new large runner (#8134)
**Context:**
Change the large runners that pennylane uses.

This would swap any instance of GitHub based large runners to the same
large runners offered by a third-part vendor
[blacksmith](https://blacksmith.sh).

**Description of the Change:**
This pull request changes the name of the runner returned when a large
runner is requested.

Which affects:
- All tests run on `master` branch
- All tests run within the merge queue
- All tests run on PRs that have the `urgent` label

**Benefits:**

- Lower costs, blacksmith offers identical environment setup to GitHub
large runners but at half the price per minute. So even without
performance benefits, we slash monthly costs for CI in half.
- Faster vcpu clock cycle and higher single thread performance. From my
coarse testing, the entire test suite has had a 30% speed-up. Tests that
were running in the 18-20 min range ran in 10-12 min. Tests that ran in
12-15min ran in 6-9 min.
- Faster runner boot up. A substantial speed up for urgent PRs waiting
for a runner.
- They also offer a dashboard for all jobs run on their runners. More
details regarding this coming soon.

**Possible Drawbacks:**

Though they are supposed to be a full drop-in replacement of the GitHub
4vcpu runner we are using, we had one test fail. I had to reduce the
feasibility tolerance to get the test to pass.

It is unclear why this happened, unless the the higher clock cycle
somehow changes the behaviour cvxpy.

UPDATE: Though a lot of debugging was performed, I could not find a root
cause on why
`tests/kernels/test_kernels.py::TestRegularization.test_closest_psd_matrix_small_perturb`
was failing.

However, the tolerance strictness on the test did make it unstable, for
example, in Python 3.11, it fails on the new runners I was testing, but
in Python 3.12, it fails in GitHub runners as well. Reducing the
tolerance from `1e-10` to `2e-10` made it pass across Python versions
and across different hardware.


**Related GitHub Issues:**
None.