Integrate dynamic wires with default qubit (#7718)
**Context:**

We've added instructions for using locally scoped wires in a pennylane
program and a transform for assigning those to the global wire order,
but it is not yet integrated.

**Description of the Change:**

Adds a `device_resolve_dynamic_wires` that takes the device wires and
applies `resolve_dynamic_wires` with the correct `min_int` and `zeroed`
kwargs.

**Benefits:**

Locally scoped wires can be used without manually applying a resolution
transform.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-95769]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>