chore: update `autograph` documentation to capture new capabilities (#8132)
**Context:**

Recent work went into getting PennyLane's autograph module up to spec
with Catalyst. The main highlights are: logical operation support,
pythonic index assignment and pythonic array updates.

**Description of the Change:**

Updates the documentation to remove previous sharp bits that now work.

ℹ️ Please note that I am not adding the features here as I find it weird
to let the user know that standard pythonic index assignment works,
instead it is something the user expects to work and so it is not a
sharp bit / feature.

**Benefits:** Accurate guide.

[sc-97972]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>