Revert xfailed/skipped tests in unified compiler module (#8178)
**Context:**
In https://github.com/PennyLaneAI/pennylane/pull/8161, for the sake of
the robustness of our CI, we temporarily turn on and off the capture
mode. However, for one thing this is a temp downgrading that altered
what those tests expect to check originally, for another it is not a
good practice to play magic over the global env manager. Therefore, we
need to revert the temp fix ASAP once Catalyst has been updated with
latest version of Lightning and PennyLane.

**Description of the Change:**
As mentioned, revert the affected tests in xdsl tests.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-98497]