Add xDSL pass to decompose `mbqc.graph_state_prep` to quantum ops (#8090)
**Context:** We added the `mbqc.graph_state_prep` op to the Python
compiler in #8059 to represent the allocation of a new graph state for
MBQC workloads. However, programs containing this op in the IR cannot be
executed on simulators because we had no lowering patterns for it.

**Description of the Change:** This PR adds two new xDSL passes to the
Python compiler:

1. `decompose-graph-state`, which decomposes `mbqc.graph_state_prep`
operations to their corresponding set of quantum operations for
execution on state simulators, and
2. `null-decompose-graph-state`, which replaces `mbqc.graph_state_prep`
operations with single quantum-register allocation operations for
execution on null devices.

**Example:** Consider the following two-qubit graph state with the
following qubit connectivity:

```
0 -- 1
```

This graph has the adjacency matrix representation

```
0  1
1  0
```

and densely packed adjacency matrix representation (which is the
lower-triangular part of the matrix, excluding the diagonal elements,
stored in a flat array)

```
[1]
```

Given the input IR,

```mlir
func.func @circuit() {
    %adj_matrix = arith.constant dense<[1]> : tensor<1xi1>
    %qreg = mbqc.graph_state_prep (%adj_matrix : tensor<1xi1>) [init "Hadamard", entangle "CZ"] : !quantum.reg
    func.return
}
```

the resulting IR after the `decompose-graph-state` pass is:

```mlir
func.func @circuit() {
    %0 = quantum.alloc(2) : !quantum.reg
    %1 = quantum.extract %0[0] : !quantum.reg -> !quantum.bit
    %2 = quantum.extract %0[1] : !quantum.reg -> !quantum.bit
    %3 = quantum.custom "Hadamard"() %1 : !quantum.bit
    %4 = quantum.custom "Hadamard"() %2 : !quantum.bit
    %5, %6 = quantum.custom "CZ"() %3, %4 : !quantum.bit, !quantum.bit
    %7 = quantum.insert %0[0], %5 : !quantum.reg, !quantum.bit
    %qreg = quantum.insert %7[1], %6 : !quantum.reg, !quantum.bit
    func.return
}
``` 

and after the `null-decompose-graph-state`:

```mlir
func.func @circuit() {
    %qreg = quantum.alloc(2) : !quantum.reg
    func.return
}
```

**Benefits:** This change makes it possible to validate current and
future compilation passes for MBQC workloads on state simulators and
null devices.

**Possible Drawbacks:** As with most MBQC development work, the APIs and
usage patterns of this compilation pass are subject to change without
notice.

[sc-97316]